Rename ElasticExtensions to OpenSearchExtensions
