+++ alerting/src/main/kotlin/org/opensearch/alerting/model/ActionExecutionResult.kt
@@ -5,8 +5,8 @@
 
 package org.opensearch.alerting.model
 
-import org.opensearch.alerting.elasticapi.instant
-import org.opensearch.alerting.elasticapi.optionalTimeField
+import org.opensearch.alerting.opensearchapi.instant
+import org.opensearch.alerting.opensearchapi.optionalTimeField
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
 import org.opensearch.common.io.stream.Writeable
