+++ alerting/src/main/kotlin/org/opensearch/alerting/model/destination/Destination.kt
@@ -12,10 +12,10 @@ import org.opensearch.alerting.destination.message.ChimeMessage
 import org.opensearch.alerting.destination.message.CustomWebhookMessage
 import org.opensearch.alerting.destination.message.EmailMessage
 import org.opensearch.alerting.destination.message.SlackMessage
-import org.opensearch.alerting.elasticapi.convertToMap
-import org.opensearch.alerting.elasticapi.instant
-import org.opensearch.alerting.elasticapi.optionalTimeField
-import org.opensearch.alerting.elasticapi.optionalUserField
+import org.opensearch.alerting.opensearchapi.convertToMap
+import org.opensearch.alerting.opensearchapi.instant
+import org.opensearch.alerting.opensearchapi.optionalTimeField
+import org.opensearch.alerting.opensearchapi.optionalUserField
 import org.opensearch.alerting.model.destination.email.Email
 import org.opensearch.alerting.util.DestinationType
 import org.opensearch.alerting.util.IndexUtils.Companion.NO_SCHEMA_VERSION
