+++ alerting/src/main/kotlin/org/opensearch/alerting/model/destination/Slack.kt
@@ -5,7 +5,7 @@
 
 package org.opensearch.alerting.model.destination
 
-import org.opensearch.alerting.elasticapi.string
+import org.opensearch.alerting.opensearchapi.string
 import org.opensearch.common.Strings
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
