+++ alerting/src/test/kotlin/org/opensearch/alerting/model/XContentTests.kt
@@ -7,7 +7,7 @@ package org.opensearch.alerting.model
 
 import org.opensearch.alerting.builder
 import org.opensearch.alerting.core.model.SearchInput
-import org.opensearch.alerting.elasticapi.string
+import org.opensearch.alerting.opensearchapi.string
 import org.opensearch.alerting.model.action.Action
 import org.opensearch.alerting.model.action.ActionExecutionPolicy
 import org.opensearch.alerting.model.action.PerExecutionActionScope
